[JSI][iOS] Run the Task encoding tests on a runtime with a JavaScript thread
`Task.encode` settles its promise from a detached task on the cooperative pool. The tests gave it a
standalone `JavaScriptRuntime()`, whose scheduler has no JavaScript thread and runs scheduled work
inline on the calling thread, so the settle executed JavaScript concurrently with the test body on the
same Hermes runtime. The first test never awaited the promise at all. The race surfaced on CI as
`hermes_fatal` from the interpreter on a cooperative thread, and locally as an occasional exit 65 with
a truncated test count.

`TestRuntimeScheduler`, previously private to the runtime tests, moves to `Tests/Support` with two
helpers: `makeRuntime()` builds a non-owning runtime that dispatches scheduled work to the scheduler's
thread, and `runIsolated` runs actor-isolated work there. The Task tests create the promise through
`runIsolated`, so the settle lands on the same thread, and `await()` stays a plain Swift suspension
that resumes on the settling thread.
